Mark refreshToken in TokenPair as `@NonNull`
Previously the refreshToken passed in TokenPair (`@Nullable`) was passed
back as `@NonNull` in authenticate() which Kotlin non-null checks didn't
like. If the refresh token really can't be used, then an e.g. empty
string can be used instead. Marking the refreshToken in authenticate()
as `@Nullable` would increase the complexity when the refreshToken is
always non-null.